+++ test/cell_multi_cell_refactor/__init__.py
@@ -1,1 +1,0 @@
-"""these tests pertain to the page breaking refactor of cell and multicell"""