+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 
 ## [2.2.1] - Not released yet
 ### Added
-- `FPDF.image` now accept also a `Pillow.Image.Image` as input
+- `FPDF.image` now accepts also a `Pillow.Image.Image` as input
+- `FPDF.multi_cell` parameters evolve in order to generate tables with multiline text in cells:
+  * its `ln` parameter now accepts a value of `3` that sets the new position to the right without altering vertical offset
+  * a new optional `max_line_height` parameter sets a maximum height of each sub-cell generated
 - documentation on how to add content to existing PDFs
 ### Fixed
-- `alias_nb_pages`: fixed this feature that was broken, probably since v2.0.6
-- `set_font`: fixed a bug where calling it several times, with & without the same parameters,
+- `FPDF.alias_nb_pages`: fixed this feature that was broken since v2.0.6
+- `FPDF.set_font`: fixed a bug where calling it several times, with & without the same parameters,
 prevented strings passed first to the text-rendering methods to be displayed.
 ### Deprecated
 - the `dest` parameter of `FPDF.output` method